+++ lib/src/fluttercrew_singletons.dart
@@ -1,3 +1,4 @@
 library fluttercrew.singletons;
 
+export 'singletons/ui/ui.dart';
 export 'singletons/utils/utils.dart';