+++ lib/src/singletons/ui/ui.dart
@@ -1,19 +1,24 @@
 import 'dart:async';
 
+export 'widgets/splash_screen.dart';
+
+/// UI singleton
 class UI {
+  /// UI singleton
   factory UI() => _instance;
   UI._privateConstructor();
   static final UI _instance = UI._privateConstructor();
 
-  StreamController<bool> _splashController;
+  final StreamController<bool> _splashController = StreamController<bool>();
+
+  /// stream of the splash widget
   Stream<bool> get splashStream => _splashController.stream;
-  Function(bool) get setSplash => _splashController.add;
 
-  Future<void> init() async {
-    _splashController ??= StreamController<bool>();
-  }
+  /// function to add events to splash widget
+  void Function(bool) get setSplash => _splashController.add;
 
+  /// dispose the stream controller
   Future<void> dispose() async {
-    await _splashController?.close();
+    await _splashController.close();
   }
 }