+++ lib/src/singletons/utils/utils.dart
@@ -3,19 +3,25 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import 'utils/custom_logger.dart';
 
+/// [Utils] singleton
 class Utils {
+  /// [Utils] singleton
   factory Utils() => _instance;
   Utils._privateConstructor();
   static final Utils _instance = Utils._privateConstructor();
 
-  SharedPreferences _prefs;
+  late SharedPreferences _prefs;
+  late Logger _logger;
+
+  /// [SharedPreferences] instance
   SharedPreferences get prefs => _prefs;
 
-  Logger _logger;
+  /// [Logger] instance
   Logger get logger => _logger;
 
-  Future<void> init({String projectName}) async {
-    _prefs ??= await SharedPreferences.getInstance();
-    _logger ??= customLogger(projectName);
+  /// initialize [Utils] singleton
+  Future<void> init({String projectName = ''}) async {
+    _prefs = await SharedPreferences.getInstance();
+    _logger = customLogger(projectName);
   }
 }