+++ CHANGELOG.md
@@ -5,3 +5,39 @@
 ## [0.0.2] - TODO: Add Utils.
 
 * Added utils.
+
+## [0.0.3] - TODO: Added Utils.
+
+* Added utils.
+
+## [0.0.4] - TODO: Added UI.
+
+* Added UI.
+
+## [0.0.5] - TODO: Added Splash Screen.
+
+* Added Splash Screen.
+
+## [0.0.6] - TODO: Added Splash Screen duration.
+
+* Added Splash Screen duration.
+
+## [0.0.7] - TODO: Added Splash Screen duration.
+
+* Added Splash Screen delay, and show/hide durations.
+
+## [0.0.8] - TODO: Fixed Splash Screen.
+
+* Fixed Splash Screen mounted bug.
+
+## [0.0.9] - TODO: Switch splash to AnimatedCrossFade.
+
+* Switched Splash Screen to AnimatedCrossFade.
+
+## [0.1.0] - TODO: null safety.
+
+* Added null safety.
+
+## [0.1.1] - TODO: null safety.
+
+* Fixed bugs.